refactor(web): split cloud-tasks.tsx 502->132 LOC (Cluster A2 - final)
WHY: cloud-tasks.tsx packed status/agent display config, a row mapper, a
relative-time formatter, a single-pass stats tally, two badges, an expandable
row, and the realtime data layer into 502 lines - over the 400-LOC ceiling,
with the tally + formatter + mapper untestable while embedded. Completes
Cluster A2 (all 8 oversized components split).

Extracted into a co-located cloud-tasks/ directory:
- task-format.ts  - STATUS_CONFIG/AGENT_COLORS + formatRelativeTime +
                   computeTaskStats (pure single-pass O(n) tally); now testable
- rowToTask.ts    - snake->camel DB row mapper (pure)
- StatusBadge.tsx / AgentBadge.tsx / TaskRow.tsx - presentational
- useCloudTasks.ts - load + realtime subscription + optimistic cancel
- types.ts        - CloudTasksPanelProps
- __tests__/task-format.test.ts - 8 tests pinning the stats tally + formatter +
                   mapper

The orchestrator (132 LOC) re-exports CloudTasksPanelProps + keeps the named
CloudTasksPanel export, so the dynamic() import in dashboard-realtime.tsx and the
perf-regression bundle-ratchet guard are unchanged. cloud-tasks.tsx (file)
resolves before cloud-tasks/ (dir) - verified by a clean next build.

Verification: web typecheck + lint clean; full suite 3918 pass (+8 new);
perf-regression dynamic-import guard green; next build exit 0.
